Updated routes and api end points
