+++ backend/auth.js
@@ -39,7 +39,6 @@ export const protect = (req, res, next) => {
 
   try {
     const user = jwt.verify(token, process.env.JWT_SECRET);
-    console.log("Verified User:", user, process.env.JWT_SECRET);
     req.user = user;
     next();
   } catch (e) {
